+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - registration functionality ([105](https://github.com/boarlabsxyz/razom/pull/105))
 - regionsList component ([117](https://github.com/boarlabsxyz/razom/pull/117))
+- reset filter button to the map section ([128](https://github.com/boarlabsxyz/razom/pull/128))
 
 ### Fixed